refactor: create reusable FormObserver
