+++ packages/client/src/pages/report/components/ReportFormObserver.tsx
@@ -1,0 +1,7 @@
+import { FC, useEffect, useRef } from 'react';
+
+interface ReportFormObserverProps {}
+
+export const ReportFormObserver: FC<ReportFormObserverProps> = ({}) => {
+  return null;
+};
